Separate JDBC data source config from jOOQ config.
